Update test files

+++ adl/test/scenarios/v2/multiple/apimanagement/output/anonymous/alias_384.ts
@@ -1,1 +1,1 @@
-export type alias_384 = double;
+export type alias_384 = int32;
