+++ adl/test/scenarios/v2/multiple/apimanagement/output/anonymous/alias_411.ts
@@ -1,1 +1,1 @@
-export type alias_411 = int64;
+export type alias_411 = int32;
