+++ adl/test/scenarios/v2/multiple/apimanagement/output/anonymous/alias_412.ts
@@ -1,1 +1,1 @@
-export type alias_412 = int32;
+export type alias_412 = int64;
